requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
PyMuPDF
google-generativeai
python-dotenv
